clear most warning

+++ Physics/Source/GrenadeScript.cpp
@@ -2,7 +2,6 @@
 #include "EntityScript.h"
 #include "AdvancedParticleSpawnerScript.h"
 
-
 GrenadeScript::GrenadeScript(float Lifespan, float Damage, float ExplosionRadius)
 	: ProjectileScript(Lifespan, Damage)
 	, m_fExplosionRadius(ExplosionRadius)
@@ -12,7 +11,6 @@ GrenadeScript::GrenadeScript(float Lifespan, float Damage, float ExplosionRadius
 	m_iGrenadeCount = 5;
 	m_fExplosionDamage = m_fDamage * 2;
 }
-
 
 GrenadeScript::~GrenadeScript()
 {
@@ -47,16 +45,16 @@ void GrenadeScript::Explode(void)
 
 void GrenadeScript::Collide(GameObject* go)
 {
-	EntityScript* es = go->GetComponent<EntityScript>(true);	
-	
+	EntityScript* es = go->GetComponent<EntityScript>(true);
+
 	if (es)
 	{
-		es->Damage(m_fDamage);
+		es->Damage((int)m_fDamage);
 		Vector3 relDir = go->TRANS->GetPosition() - GetPosition();
 
 		Explode();
 
-		es->Damage(m_fExplosionDamage);
+		es->Damage((int)m_fExplosionDamage);
 		go->RIGID->AddForce(relDir.Normalize() * 50);
 	}
 }
